Make it work
NOTE:
I finally understood that an interpreter *interprets* things.
And now I can make sense of what I'm writing and not just random guessing
